kmrl-system/backend/kmrl-gateway/requirements_enhanced.txt
# Enhanced KMRL Gateway Requirements
# PostgreSQL-integrated document processing with real-time updates

# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database and ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Task queue and async processing
celery>=5.3.0
redis>=5.0.0

# File storage and processing
boto3>=1.34.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
Pillow>=10.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Environment and configuration
python-dotenv>=1.0.0

# WebSocket support
websockets>=12.0

# File validation and security
python-magic>=0.4.27

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
